+++ R/list_vpts_aloft.R
@@ -28,8 +28,8 @@ list_vpts_aloft <- function(date_min = NULL,
                             format = "csv", # also hdf5
                             source = "baltrad", # also ecog-04003
                             show_warnings = TRUE) {
-  warning("`list_vpts_aloft()` will be deprecated in a future version. See `getRad::get_vpts()` and `getRad::aloft_data_coverage()` instead.",
-          call. = FALSE)
+
+  lifecycle::deprecate_warn("2025", "list_vpts_aloft()", "getRad::get_vpts()")
 
   # Check if aws.s3 is installed
   # NOTE added because aws.s3 is schedueled to be moved to Suggests